frontend: refactor mentors
